Move wait_obj and chan_plugin to libfreerdputils; remove libcommon.
