+++ channels/drdynvc/Makefile.am
@@ -10,12 +10,12 @@ drdynvc_la_SOURCES = \
 	drdynvc_types.h \
 	drdynvc_main.c drdynvc_main.h
 
-drdynvc_la_CFLAGS = -I$(top_srcdir)/include -I$(srcdir)/../common \
+drdynvc_la_CFLAGS = -I$(top_srcdir)/include \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 drdynvc_la_LDFLAGS = -avoid-version -module
 
-drdynvc_la_LIBADD = ../common/libcommon.la
+drdynvc_la_LIBADD = ../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
