+++ channels/drdynvc/audin/pulse/Makefile.am
@@ -9,12 +9,12 @@ audin_pulse_la_SOURCES = \
 	audin_pulse.c
 
 audin_pulse_la_CFLAGS = @PULSE_CFLAGS@ \
-	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. -I$(srcdir)/../../../common \
+	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 audin_pulse_la_LDFLAGS = -avoid-version -module
 
-audin_pulse_la_LIBADD = @PULSE_LIBS@ ../../../common/libcommon.la
+audin_pulse_la_LIBADD = @PULSE_LIBS@ ../../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
