+++ channels/drdynvc/drdynvc_types.h
@@ -24,7 +24,6 @@
 #include <freerdp/vchan.h>
 #include <freerdp/dvc.h>
 #include <freerdp/utils.h>
-#include "chan_plugin.h"
 
 #define LOG_LEVEL 1
 #define LLOG(_level, _args) \
