+++ channels/drdynvc/pnpdr/Makefile.am
@@ -8,12 +8,12 @@ pnpdr_LTLIBRARIES = pnpdr.la
 pnpdr_la_SOURCES = \
 	pnpdr_main.c
 
-pnpdr_la_CFLAGS = -I$(top_srcdir)/include -I$(srcdir)/../../common \
+pnpdr_la_CFLAGS = -I$(top_srcdir)/include \
 	-I$(srcdir)/.. -DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 pnpdr_la_LDFLAGS = -avoid-version -module
 
-pnpdr_la_LIBADD = ../../common/libcommon.la
+pnpdr_la_LIBADD = ../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
