+++ channels/drdynvc/tsmf/alsa/Makefile.am
@@ -9,12 +9,12 @@ tsmf_alsa_la_SOURCES = \
 	tsmf_alsa.c
 
 tsmf_alsa_la_CFLAGS = @ALSA_CFLAGS@ \
-	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. -I$(srcdir)/../../../common \
+	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 tsmf_alsa_la_LDFLAGS = -avoid-version -module
 
-tsmf_alsa_la_LIBADD = @ALSA_LIBS@ ../../../common/libcommon.la
+tsmf_alsa_la_LIBADD = @ALSA_LIBS@ ../../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
