+++ channels/drdynvc/tsmf/ffmpeg/Makefile.am
@@ -9,12 +9,12 @@ tsmf_ffmpeg_la_SOURCES = \
 	tsmf_ffmpeg.c
 
 tsmf_ffmpeg_la_CFLAGS = @FFMPEG_CFLAGS@ \
-	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. -I$(srcdir)/../../../common \
+	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 tsmf_ffmpeg_la_LDFLAGS = -avoid-version -module
 
-tsmf_ffmpeg_la_LIBADD = @FFMPEG_LIBS@ ../../../common/libcommon.la
+tsmf_ffmpeg_la_LIBADD = @FFMPEG_LIBS@ ../../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
