+++ channels/drdynvc/tsmf/pulse/Makefile.am
@@ -9,12 +9,12 @@ tsmf_pulse_la_SOURCES = \
 	tsmf_pulse.c
 
 tsmf_pulse_la_CFLAGS = @PULSE_CFLAGS@ \
-	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. -I$(srcdir)/../../../common \
+	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../.. \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 tsmf_pulse_la_LDFLAGS = -avoid-version -module
 
-tsmf_pulse_la_LIBADD = @PULSE_LIBS@ ../../../common/libcommon.la
+tsmf_pulse_la_LIBADD = @PULSE_LIBS@ ../../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
