+++ channels/rdpdbg/Makefile.am
@@ -8,11 +8,11 @@ rdpdbg_LTLIBRARIES = rdpdbg.la
 rdpdbg_la_SOURCES = \
 	rdpdbg_main.c rdpdbg_main.h
 
-rdpdbg_la_CFLAGS = -I$(top_srcdir)/include -I$(srcdir)/../common
+rdpdbg_la_CFLAGS = -I$(top_srcdir)/include 
 
 rdpdbg_la_LDFLAGS = -avoid-version -module
 
-rdpdbg_la_LIBADD = ../common/libcommon.la
+rdpdbg_la_LIBADD = ../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
