+++ channels/rdpdbg/rdpdbg_main.c
@@ -25,8 +25,6 @@
 #include <freerdp/types_ui.h>
 #include <freerdp/vchan.h>
 #include <freerdp/utils.h>
-#include "chan_plugin.h"
-#include "wait_obj.h"
 #include "rdpdbg_main.h"
 
 #define LOG_LEVEL 11
