+++ channels/rdpdr/disk/Makefile.am
@@ -8,12 +8,12 @@ disk_LTLIBRARIES = disk.la
 disk_la_SOURCES = \
 	disk_main.c
 
-disk_la_CFLAGS = -I$(top_srcdir)/include -I$(srcdir)/../../common \
+disk_la_CFLAGS = -I$(top_srcdir)/include \
 	-I$(srcdir)/.. -DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 disk_la_LDFLAGS = -avoid-version -module
 
-disk_la_LIBADD = ../../common/libcommon.la
+disk_la_LIBADD = ../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
