+++ channels/rdpdr/printer/Makefile.am
@@ -8,12 +8,12 @@ printer_LTLIBRARIES = printer.la
 printer_la_SOURCES = \
 	printer_main.c printer_main.h
 
-printer_la_CFLAGS = -I$(top_srcdir)/include -I$(srcdir)/../../common \
+printer_la_CFLAGS = -I$(top_srcdir)/include \
 	-I$(srcdir)/.. -DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 printer_la_LDFLAGS = -avoid-version -module
 
-printer_la_LIBADD = ../../common/libcommon.la
+printer_la_LIBADD = ../../../libfreerdputils/libfreerdputils.la
 
 if RDPDR_PRINTER_CUPS
 printer_la_SOURCES += printer_cups.c
