+++ channels/rdpdr/rdpdr_types.h
@@ -23,7 +23,6 @@
 #include <freerdp/types_ui.h>
 #include <freerdp/vchan.h>
 #include <freerdp/utils.h>
-#include "chan_plugin.h"
 
 #define LOG_LEVEL 1
 #define LLOG(_level, _args) \
