+++ channels/rdpsnd/alsa/Makefile.am
@@ -9,12 +9,12 @@ rdpsnd_alsa_la_SOURCES = \
 	rdpsnd_alsa.c
 
 rdpsnd_alsa_la_CFLAGS = @ALSA_CFLAGS@ \
-	-I$(top_srcdir)/include -I$(srcdir)/.. -I$(srcdir)/../../common \
+	-I$(top_srcdir)/include -I$(srcdir)/.. \
 	-DPLUGIN_PATH=\"$(PLUGIN_PATH)\"
 
 rdpsnd_alsa_la_LDFLAGS = -avoid-version -module
 
-rdpsnd_alsa_la_LIBADD = @ALSA_LIBS@
+rdpsnd_alsa_la_LIBADD = @ALSA_LIBS@ ../../../libfreerdputils/libfreerdputils.la
 
 # extra
 EXTRA_DIST =
