+++ configure.ac
@@ -1058,7 +1058,6 @@ libfreerdpchanman/Makefile
 X11/Makefile
 dfb/Makefile
 win/Makefile
-channels/common/Makefile
 channels/rdpdbg/Makefile
 channels/rdpsnd/Makefile
 channels/rdpsnd/alsa/Makefile
